Move all the files to the top level dir
