+++ arrayToMap.js
