+++ capitalizeFirstChar.js
