+++ getHostNameFromUrl.js
