+++ getQueryParams.js
