+++ getQueryString.js
