+++ loadExternalScripts.js
