+++ swapArrayItems.js
